+++ dart/geobase/lib/src/vector_data/model/geometry/geometry_collection.dart
@@ -19,6 +19,7 @@ import '/src/coordinates/base/position_scheme.dart';
 import '/src/coordinates/projection/projection.dart';
 import '/src/utils/bounded_utils.dart';
 import '/src/utils/coord_type.dart';
+import '/src/utils/geometry_calculations_cartesian.dart';
 import '/src/vector/content/geometry_content.dart';
 import '/src/vector/encoding/binary_format.dart';
 import '/src/vector/encoding/text_format.dart';
@@ -356,6 +357,22 @@ class GeometryCollection<E extends Geometry> extends Geometry {
       area += geom.area2D();
     }
     return area;
+  }
+
+  @override
+  Position? centroid2D() {
+    final calculator = CompositeCentroid();
+    for (final geom in geometries) {
+      final centroid = geom.centroid2D();
+      if (centroid != null) {
+        calculator.addCentroid2D(
+          centroid,
+          area: geom.area2D(),
+          length: geom.length2D(),
+        );
+      }
+    }
+    return calculator.centroid();
   }
 
   @override