+++ dart/geobase/lib/src/vector_data/model/geometry/multi_linestring.dart
@@ -9,6 +9,7 @@
 import 'dart:typed_data';
 
 import '/src/common/codes/coords.dart';
+import '/src/common/codes/dimensionality.dart';
 import '/src/common/codes/geom.dart';
 import '/src/common/constants/epsilon.dart';
 import '/src/common/reference/coord_ref_sys.dart';
@@ -21,6 +22,7 @@ import '/src/coordinates/projection/projection.dart';
 import '/src/utils/bounded_utils.dart';
 import '/src/utils/coord_positions.dart';
 import '/src/utils/coord_type.dart';
+import '/src/utils/geometry_calculations_cartesian.dart';
 import '/src/vector/content/simple_geometry_content.dart';
 import '/src/vector/encoding/binary_format.dart';
 import '/src/vector/encoding/text_format.dart';
@@ -467,6 +469,21 @@ class MultiLineString extends SimpleGeometry {
 
   @override
   double area2D() => 0.0;
+
+  @override
+  Position? centroid2D() {
+    final calculator = CompositeCentroid();
+    for (final chain in chains) {
+      final centroid = chain.centroid2D(dimensionality: Dimensionality.linear);
+      if (centroid != null) {
+        calculator.addCentroid2D(
+          centroid,
+          length: chain.length2D(),
+        );
+      }
+    }
+    return calculator.centroid();
+  }
 
   @override
   void writeTo(SimpleGeometryContent writer, {String? name}) =>