+++ dart/geobase/CHANGELOG.md
@@ -10,6 +10,7 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
 * [Add EWKB encoding and decoding support on geobase #165](https://github.com/navibyte/geospatial/issues/165)
 * [Add EWKT support](https://github.com/navibyte/geospatial/issues/29)
 * [Support for GeoJSON Text Sequences](https://github.com/navibyte/geospatial/issues/217)
+* [Centroid for all geometry objects #223](https://github.com/navibyte/geospatial/issues/223)
 
 ## 1.0.2